Fixed Expected Long instead of Float Error

+++ Google_Colab-Interactive_AI_Playground/LLM_Small1x1.py
@@ -282,7 +282,7 @@ class GPTModel(nn.Module):
 
     def forward(self, in_idx):
         batch_size, seq_len = in_idx.shape
-        in_idx.to(device)
+        in_idx.to(device).long()
         tok_embeds = self.tok_emb(in_idx)
         pos_embeds = self.pos_emb(torch.arange(seq_len, device=in_idx.device))
         x = tok_embeds + pos_embeds  # Shape [batch_size, num_tokens, emb_size]
